add proxies and add test
