+++ python-package/cartiflette/cartiflette/utils.py
@@ -1,7 +1,5 @@
 import typing
 import logging
-from datetime import date
-import geopandas as gpd
 
 from cartiflette.constants import BUCKET, PATH_WITHIN_BUCKET
 
@@ -154,58 +152,3 @@ def create_path_bucket(config: ConfigDict) -> str:
         write_path += f"/raw.{vectorfile_format}"
 
     return write_path
-
-
-def download_cartiflette_single(
-    *args,
-    bucket: str = BUCKET,
-    path_within_bucket: str = PATH_WITHIN_BUCKET,
-    provider: str = "IGN",
-    dataset_family: str = "ADMINEXPRESS",
-    source: str = "EXPRESS-COG-TERRITOIRE",
-    vectorfile_format: str = "geojson",
-    borders: str = "COMMUNE",
-    filter_by: str = "region",
-    territory: str = "metropole",
-    year: typing.Union[str, int, float] = None,
-    value: typing.Union[str, int, float] = "28",
-    crs: typing.Union[list, str, int, float] = 2154,
-    simplification: typing.Union[str, int, float] = None,
-    filename: str = "raw",
-    **kwargs,
-):
-    if not year:
-        year = str(date.today().year)
-
-    corresp_filter_by_columns, format_read, driver = standardize_inputs(
-        vectorfile_format
-    )
-
-    url = create_path_bucket(
-        {
-            "bucket": bucket,
-            "path_within_bucket": path_within_bucket,
-            "vectorfile_format": format_read,
-            "territory": territory,
-            "borders": borders,
-            "filter_by": filter_by,
-            "year": year,
-            "value": value,
-            "crs": crs,
-            "provider": provider,
-            "dataset_family": dataset_family,
-            "source": source,
-            "simplification": simplification,
-            "filename": filename,
-        }
-    )
-
-    url = f"https://minio.lab.sspcloud.fr/{url}"
-
-    try:
-        gdf = gpd.read_file(url)
-    except Exception as e:
-        logger.error(f"There was an error while reading the file from the URL: {url}")
-        logger.error(f"Error message: {str(e)}")
-    else:
-        return gdf
